requirements.txt
flask==2.0.1
ibm-watson==6.1.0
python-dotenv==0.19.0
pytest==7.0.1
flake8==4.0.1
